+++ data/TaskList.txt
@@ -4,3 +4,7 @@ T | 0 | add
 T | 0 | add
 T | 1 | sleep
 E | 0 | sleep | 23 Aug 2023 6:00PM | 23 Aug 2023 6:00PM
+T | 0 | add
+T | 0 | sleep
+D | 0 | sleep | 23 Aug 2023 8:00PM
+T | 0 | add
