+++ src/main/java/Comm/AddDeadlineCommand.java
@@ -1,3 +1,8 @@
+package Comm;
+import TaskManager.Deadlines;
+import Ui.Ui;
+import Storage.TaskList;
+import Storage.FileHandler;
 public class AddDeadlineCommand extends Command{
     private String userInput;
     private String dueDateStr;
