+++ src/main/java/Comm/ByeCommand.java
@@ -1,3 +1,7 @@
+package Comm;
+import Ui.Ui;
+import Storage.TaskList;
+import Storage.FileHandler;
 public class ByeCommand extends Command{
 
     public ByeCommand(){
