+++ src/main/java/Comm/Command.java
@@ -1,4 +1,8 @@
+package Comm;
 import java.io.File;
+import Ui.Ui;
+import Storage.TaskList;
+import Storage.FileHandler;
 
 public abstract class Command {
