+++ src/main/java/Comm/DeleteCommand.java
@@ -1,8 +1,13 @@
-public class UnmarkCommand extends Command{
+package Comm;
+import Ui.Ui;
+import Storage.TaskList;
+import Storage.FileHandler;
+import TaskManager.Tasks;
+public class DeleteCommand extends Command{
 
     private String userInput;
 
-    public UnmarkCommand(String userInput) {
+    public DeleteCommand(String userInput) {
         this.userInput = userInput;
     }
 
@@ -10,9 +15,10 @@ public class UnmarkCommand extends Command{
     public void execute(TaskList t, Ui ui, FileHandler f) {
         try {
             int index = Integer.parseInt(userInput.substring(7));
-            t.get(index - 1).markNotDone();
+            Tasks deleted = t.get(index - 1);
+            t.remove(index - 1);
             FileHandler.writeTasksToFile(t);
-            ui.unmark(index);
+            ui.delete(deleted);
         } catch (IndexOutOfBoundsException e) {
             ui.IOOBExceptionMessage();
         }
