+++ src/main/java/Comm/ErrorCommand.java
@@ -1,11 +1,16 @@
-public class ListCommand extends Command{
+package Comm;
+import Ui.Ui;
+import Storage.TaskList;
+import Storage.FileHandler;
 
-    public ListCommand() {
+public class ErrorCommand extends Command{
+
+    public ErrorCommand(){
     }
 
     @Override
     public void execute(TaskList t, Ui ui, FileHandler f) {
-        ui.List();
+
     }
 
     @Override
