+++ src/main/java/Comm/MiscCommand.java
@@ -1,3 +1,8 @@
+package Comm;
+import Ui.Ui;
+import Storage.TaskList;
+import Storage.FileHandler;
+
 public class MiscCommand extends Command{
 
     public MiscCommand(){
