+++ src/main/java/Duke.java
@@ -1,4 +1,8 @@
-import java.util.Scanner;
+import Parser.Parser;
+import Storage.FileHandler;
+import Storage.TaskList;
+import Ui.Ui;
+import Comm.Command;
 
 public class Duke {
     private FileHandler fileHandler;
