--- src/main/java/ErrorCommand.java
@@ -1,15 +1,0 @@
-public class ErrorCommand extends Command{
-
-    public ErrorCommand(){
-    }
-
-    @Override
-    public void execute(TaskList t, Ui ui, FileHandler f) {
-
-    }
-
-    @Override
-    public boolean isExit() {
-        return false;
-    }
-}
