+++ src/main/java/TaskManager/Events.java
@@ -1,3 +1,4 @@
+package TaskManager;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 public class Events extends Tasks {
@@ -31,7 +32,6 @@ public class Events extends Tasks {
             this.taskDesc = null;
         }
     }
-
 
     public boolean isValid() {
         return taskDesc != null;
