+++ src/main/java/TaskManager/Tasks.java
@@ -1,8 +1,9 @@
+package TaskManager;
 public abstract class Tasks {
 
     boolean status;
     static int count = 0;
-    abstract String toFileString();
+    public abstract String toFileString();
     public Tasks() {
         this.status = false;
     }
